Add MSAA Option to Graphics Settings
The recently added MSAA flag option is an extra pass done after the basal AA pass. The MSAA pass greatly reduces glinting and flickering of details on ships and other models when they are far away. The MSAA settings is already shown in SCPUI as a separate dropdown list so this PR adds it as an option to KNet.

Also, this PR sets the default values for shadows to disabled and AA preset to SMAA Low, which is what FSO sets as defaults.