Update Input Settings Section
This PR disables the mouse and joystick settings, as setting these values does not do anything since those values are set in the pilot file.

All the other settings in the Settings tab change things in-game whereas this these settings did not (since it does not do anything with the command line and the settings in the pilot file take precedence). Thus this PR disables them to clean things up. Also updates the `fso` capitalization to be consistently `FSO'