Clean up similarly name open folder commands
In preparation for adding a `open screenshots` folder button I realized that many of the button function commands all use the same name (`OpenFolder`) but for different functions. I've updated the function names accordingly, and tests show things work as expected.

I understand though if you would rather not merge this PR though, as it might be not the style or cleanup ideas you had in mind.