Switch to use Compiled Bindings
This was far more invasive change that i originally belived.
Compiled Bindings are faster and they give a error during compile if there is a broken bind. BUT the variables im binding to the UI cant be private, they have to be public or internal, just like method name for button commands.
If the access level if privete they give a silent "Method.Access" error in VS console, and thats it. So petty much everything must be checked again.