Switch QtFRED and Details mini-tab positions
I know Fred and QTFred are next to each other in old Knossos, but since QTFred does not fully work it would be more useful for it to the Advanced min-tab and put Details in its place. (If QTFred ever does get done it can replace actual FRED). This change would especially help new folks who don't know what QTFred is, as they won't realize it does not fully work yet.

Happy to edit this PR further as needed, too.