fix builds from release workflow
Windows builds don't work properly when built on non-Windows platforms so
do all of the builds on a Windows runner and then switch to Linux for the
rest of the packaging and release actions.

Handle version better by pulling it from git tag. Also remove version from
AppImage filenames to ease auto-update system.