Two Fixes
The logic for `vsync` was backwards, as checking the vsync box on incorrectly set the `no_vsync` flag on. This PR fixes that, and tests confirm it not works as expected.

Also updates the user-facing wording of `cmdline` to now match the `Command Line` wording that is shown in the updated Settings tab.