Improve default visibility fro bottom buttons on quicksetup guide (#133)
On some of the text boxes with longer text I've Widens the default max width of QuickSetup Guide. I noticed I have to manually resize the window to see the buttons on the bottom to click on, and I have had a two reports of folks not knowing there were buttons to click on since they didn't see them due to the window constraints. I tested this change and it works better for always displaying the buttons. Would be happy to discuss more, too.