Fix AA name overflows and improve tool tips for video options (#158)
* Fix AA name overflows

The new text PR makes two of the longer names in the Video options slightly cutoff. One option is to move the drop downs and check boxes more to the right, but then they would be out of alignment with the rest of the columns from other sections. Another, simpler option might be to just shorten the names, since they are rather long compared to other option  names. From my tests, editing the names is the easiest and looks the best.  Would be happy to refine as requested of course.

* update tool tips to be more helpful