Remove nonsense condition
Mod compression support begins at 23.2 (specifically a 23.1 nightly), but we do have mods on knossos whose default build is 19.0, very different from 23.2.  Allowing compression to take place for mods like this does not make sense, unless the user is in Manual mode and knows that they will have to risk a newer build to correctly run the compressed modpack.  So, remove the automatic approval if the mod just happens to have a ">=" in the build version.