Update Settings headings with some slight tuning
Reporting some new testers questions, it was confusing to them that the last section had FSO Settings in the title, while none of the others did, even though they too were settings FSO uses. Thus, this PR adds tooltips to the Settings headings, and updates the name of the last item from `FSO Engine Settings` to `Specialized` (since all but one of the Settings headings are also FSO Settings).

Also I fixed a line in the `QuickSetup` that was not updated last time the Settings headings were updated.