escape underscores so that they will show on controls
Underscores are considered a special character when used in controls so they
need to be escaped using a double-underscore to force one to show up. This
adds a utility function to do the conversion and a custom data converter for
use with bindings to easily escape the underscores where needed.

(Also fixes double entries in modify engine build window)