Refresh dev fso build list during builds install/delete and add reload depedency button
The first auto updates the fso build list on the developer fso tab to account for new build install and delete.

The seconds add a reload depedency button to the mod pkg manager, this is to reload the listed versions for a depedency id so it can be updated after installing a new mod/build without reloading the entire mod editor.