Make Settings Tab always consistent
Following user testing, in instructions throughout FSO websites and discussions, the `Settings` tab in Knossos is commonly referred to. The standalone TC mode changes that name to `Options` which was b/c the name `Settings` was already used in the mod `Settings` button on the Home Screen. That button is more of an `Advanced` settings especially for TC mode, so this PR proposes changing keeping the global `Settings` tab as `Settings` so that users will always understand instructions from community members and the QuickStart Guide regarding which tab controls global settings. This PR also proposes changing the default name of the `Settings` button on the TC mode home screen to `Advanced` by default, but of course mods can set this text to whatever they want since TC mode exposes that option.

Would be happy to dicuss if useful, too, thanks!