Allow TC Mode to alter Nebula Download and Up-to-Date Text (#274)
The current TC mode options are great and really allow for a massive customization of almost everything about the home screen and launcher, except for the Nebula messages. This PR allows TC modders to set the main nebula message texts that are usually seen on the home screen, which really helps make the home screen feel fully integrated with the other customization options.

This PR also updates a few public facing capitalizations to be more consistent with other wording capitalizations.

Everything was tested and works as expected.