remove checksum generation step
GitHub will now do this for us automatically and in a more secure fashion.
These new checksums will be displayed next to each release asset as well as
being available through the REST API and gh CLI app.

Note that this feature only applies to new releases, not existing ones.

https://github.blog/changelog/2025-06-03-releases-now-expose-digests-for-release-assets/